Move ExpectTableRowCountToEqualOtherTable tests to their own file
