+++ tests/integration/data_sources_and_expectations/test_canonical_expectations.py
@@ -1,10 +1,9 @@
 from datetime import datetime, timezone
-from typing import Mapping, Sequence
+from typing import Sequence
 
 import pandas as pd
 
 import great_expectations.expectations as gxe
-from great_expectations.datasource.fluent.interfaces import Batch
 from tests.integration.conftest import parameterize_batch_for_data_sources
 from tests.integration.test_utils.data_source_config import (
     BigQueryDatasourceTestConfig,
@@ -155,57 +154,3 @@ def test_expect_column_mean_to_be_between(batch_for_datasource):
     expectation = gxe.ExpectColumnMeanToBeBetween(column="a", min_value=2, max_value=3)
     result = batch_for_datasource.validate(expectation)
     assert result.success
-
-
-class TestExpectTableRowCountToEqualOtherTable:
-    MULTI_ASSET_DATA_SOURCES: Sequence[DataSourceTestConfig] = [
-        DatabricksDatasourceTestConfig(),
-        MSSQLDatasourceTestConfig(),
-        MySQLDatasourceTestConfig(),
-        PostgreSQLDatasourceTestConfig(),
-        SnowflakeDatasourceTestConfig(),
-        SqliteDatasourceTestConfig(),
-    ]
-
-    @parameterize_batch_for_data_sources(
-        data_source_configs=MULTI_ASSET_DATA_SOURCES,
-        data=pd.DataFrame({"a": [1, 2, 3, 4]}),
-        extra_data={"other_table": pd.DataFrame({"col_b": ["a", "b", "c", "d"]})},
-    )
-    def test_success(
-        self,
-        batch_for_datasource: Batch,
-        extra_table_names_for_datasource: Mapping[str, str],
-    ):
-        other_table_name = extra_table_names_for_datasource["other_table"]
-        expectation = gxe.ExpectTableRowCountToEqualOtherTable(other_table_name=other_table_name)
-        result = batch_for_datasource.validate(expectation)
-        assert result.success
-
-    @parameterize_batch_for_data_sources(
-        data_source_configs=MULTI_ASSET_DATA_SOURCES,
-        data=pd.DataFrame({"a": [1, 2, 3, 4]}),
-        extra_data={"other_table": pd.DataFrame({"col_b": ["just_this_one!"]})},
-    )
-    def test_different_counts(
-        self,
-        batch_for_datasource: Batch,
-        extra_table_names_for_datasource: Mapping[str, str],
-    ):
-        other_table_name = extra_table_names_for_datasource["other_table"]
-        expectation = gxe.ExpectTableRowCountToEqualOtherTable(other_table_name=other_table_name)
-        result = batch_for_datasource.validate(expectation)
-        assert not result.success
-        assert result.result["observed_value"] == {
-            "self": 4,
-            "other": 1,
-        }
-
-    @parameterize_batch_for_data_sources(
-        data_source_configs=MULTI_ASSET_DATA_SOURCES,
-        data=pd.DataFrame({"a": [1, 2, 3, 4]}),
-    )
-    def test_missing_table(self, batch_for_datasource):
-        expectation = gxe.ExpectTableRowCountToEqualOtherTable(other_table_name="where_am_i")
-        result = batch_for_datasource.validate(expectation)
-        assert not result.success, "We should not find the other table, since we didn't load it."
