+++ tests/test_definitions/multi_table_expectations/expect_table_row_count_to_equal_other_table.json
@@ -6,84 +6,28 @@
         {
           "dataset_name": "expect_table_row_count_to_equal_other_table_data_1",
           "data": {
-            "c1": [
-              4,
-              5,
-              6,
-              7
-            ],
-            "c2": [
-              "a",
-              "b",
-              "c",
-              "d"
-            ],
-            "c3": [
-              null,
-              null,
-              null,
-              null
-            ],
-            "c4": [
-              4.0,
-              3.0,
-              3.5,
-              1.2
-            ]
+            "c1": [4, 5, 6, 7],
+            "c2": ["a", "b", "c", "d"],
+            "c3": [null, null, null, null],
+            "c4": [4.0, 3.0, 3.5, 1.2]
           }
         },
         {
           "dataset_name": "expect_table_row_count_to_equal_other_table_data_2",
           "data": {
-            "c1": [
-              4,
-              5,
-              6,
-              7
-            ],
-            "c2": [
-              "a",
-              "b",
-              "c",
-              "d"
-            ],
-            "c3": [
-              null,
-              null,
-              null,
-              null
-            ],
-            "c4": [
-              4.0,
-              3.0,
-              3.5,
-              1.2
-            ]
+            "c1": [4, 5, 6, 7],
+            "c2": ["a", "b", "c", "d"],
+            "c3": [null, null, null, null],
+            "c4": [4.0, 3.0, 3.5, 1.2]
           }
         },
         {
           "dataset_name": "expect_table_row_count_to_equal_other_table_data_3",
           "data": {
-            "c1": [
-              4,
-              5,
-              6
-            ],
-            "c2": [
-              "a",
-              "b",
-              "c"
-            ],
-            "c3": [
-              null,
-              null,
-              null
-            ],
-            "c4": [
-              4.0,
-              3.0,
-              3.5
-            ]
+            "c1": [4, 5, 6],
+            "c2": ["a", "b", "c"],
+            "c3": [null, null, null],
+            "c4": [4.0, 3.0, 3.5]
           }
         }
       ],
@@ -101,22 +45,6 @@
             "observed_value": {
               "self": 4,
               "other": 4
-            }
-          }
-        },
-        {
-          "title": "basic_negative",
-          "include_in_gallery": true,
-          "only_for": ["sqlalchemy"],
-          "exact_match_out": false,
-          "in": {
-            "other_table_name": "expect_table_row_count_to_equal_other_table_data_3"
-          },
-          "out": {
-            "success": false,
-            "observed_value": {
-              "self": 4,
-              "other": 3
             }
           }
         }
